+++ schema.ts
@@ -1068,9 +1068,6 @@ export type ObjectSchemaPathProperties = Pick<
   | 'userIdProperty'
   | 'groupIdProperty'
   | 'memberGroupIdProperty'
-  | 'authorityNormProperty'
-  | 'popularityNormProperty'
-  | 'popularityRankProperty'
   | 'versionProperty'
 >;
 
@@ -1143,16 +1140,6 @@ export interface IndexDefinition {
    * @hidden
    */
   popularityNormProperty?: PropertyIdentifier<string>;
-  /**
-   * @deprecated Use {@link IndexDefinition.popularityNormProperty} instead.
-   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be interpreted as a number
-   * between -1.0 and 1.0 representing the normalized popularity score of this entity compared to all other entities.
-   *
-   * Must be a {@link ValueType.Number} property.
-   * TODO(solomon): Remove this when all Packs' popularityRank is migrated to popularityNorm
-   * @hidden
-   */
-  popularityRankProperty?: PropertyIdentifier<string>;
 }
 
 /**
@@ -1402,36 +1389,6 @@ export interface ObjectSchemaDefinition<K extends string, L extends string>
    * @hidden
    */
   memberGroupIdProperty?: PropertyIdentifier<K>;
-
-  /**
-   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be interpreted as a number
-   * between -1.0 and 1.0 representing the normalized authority score of this entity compared to all other entities.
-   *
-   * Must be a {@link ValueType.Number} property.
-   * TODO(sam): Unhide this
-   * @hidden
-   */
-  authorityNormProperty?: PropertyIdentifier<K>;
-
-  /**
-   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be interpreted as a number
-   * between -1.0 and 1.0 representing the normalized popularity score of this entity compared to all other entities.
-   *
-   * Must be a {@link ValueType.Number} property.
-   * TODO(sam): Unhide this
-   * @hidden
-   */
-  popularityNormProperty?: PropertyIdentifier<K>;
-  /**
-   * @deprecated Use {@link ObjectSchemaDefinition.popularityNormProperty} instead.
-   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be interpreted as a number
-   * between -1.0 and 1.0 representing the normalized popularity score of this entity compared to all other entities.
-   *
-   * Must be a {@link ValueType.Number} property.
-   * TODO(solomon): Remove this when all Packs' popularityRank is migrated to popularityNorm
-   * @hidden
-   */
-  popularityRankProperty?: PropertyIdentifier<K>;
 
   /**
    * Defines how to index objects for use with full-text indexing.
@@ -2117,7 +2074,6 @@ function normalizeIndexDefinition(
     contextProperties,
     authorityNormProperty,
     popularityNormProperty,
-    popularityRankProperty,
     ...rest
   } = index;
   ensureNever<keyof typeof rest>();
@@ -2131,9 +2087,6 @@ function normalizeIndexDefinition(
       : undefined,
     popularityNormProperty: popularityNormProperty
       ? normalizeSchemaPropertyIdentifier(popularityNormProperty, normalizedProperties)
-      : undefined,
-    popularityRankProperty: popularityRankProperty
-      ? normalizeSchemaPropertyIdentifier(popularityRankProperty, normalizedProperties)
       : undefined,
   };
 }
@@ -2215,9 +2168,6 @@ export function normalizeObjectSchema(schema: GenericObjectSchema): GenericObjec
     userIdProperty,
     groupIdProperty,
     memberGroupIdProperty,
-    authorityNormProperty,
-    popularityNormProperty,
-    popularityRankProperty,
     versionProperty,
     index,
     parent,
@@ -2289,15 +2239,6 @@ export function normalizeObjectSchema(schema: GenericObjectSchema): GenericObjec
       : undefined,
     memberGroupIdProperty: memberGroupIdProperty
       ? normalizeSchemaPropertyIdentifier(memberGroupIdProperty, normalizedProperties)
-      : undefined,
-    authorityNormProperty: authorityNormProperty
-      ? normalizeSchemaPropertyIdentifier(authorityNormProperty, normalizedProperties)
-      : undefined,
-    popularityNormProperty: popularityNormProperty
-      ? normalizeSchemaPropertyIdentifier(popularityNormProperty, normalizedProperties)
-      : undefined,
-    popularityRankProperty: popularityRankProperty
-      ? normalizeSchemaPropertyIdentifier(popularityRankProperty, normalizedProperties)
       : undefined,
     versionProperty: versionProperty
       ? normalizeSchemaPropertyIdentifier(versionProperty, normalizedProperties)