+++ testing/upload_validation.ts
@@ -1377,7 +1377,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
     contextProperties: contextPropertiesSchema.optional(),
     authorityNormProperty: propertySchema.optional(),
     popularityNormProperty: propertySchema.optional(),
-    popularityRankProperty: propertySchema.optional(),
   });
 
   const identitySchema = zodCompleteObject<Identity>({
@@ -1424,9 +1423,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
       userEmailProperty: propertySchema.optional(),
       groupIdProperty: propertySchema.optional(),
       memberGroupIdProperty: propertySchema.optional(),
-      authorityNormProperty: propertySchema.optional(),
-      popularityNormProperty: propertySchema.optional(),
-      popularityRankProperty: propertySchema.optional(),
       versionProperty: propertySchema.optional(),
       options: zodOptionsFieldWithValues(z.object({}).passthrough(), false),
       requireForUpdates: z.boolean().optional(),
@@ -1667,30 +1663,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
           );
         };
 
-        const validateAuthorityNormProperty = () => {
-          return validateProperty(
-            'authorityNormProperty',
-            authorityNormPropertySchema => authorityNormPropertySchema.type === ValueType.Number,
-            `must refer to a "ValueType.Number" property.`,
-          );
-        };
-
-        const validatePopularityNormProperty = () => {
-          return validateProperty(
-            'popularityNormProperty',
-            popularityNormPropertySchema => popularityNormPropertySchema.type === ValueType.Number,
-            `must refer to a "ValueType.Number" property.`,
-          );
-        };
-
-        const validatePopularityRankProperty = () => {
-          return validateProperty(
-            'popularityRankProperty',
-            popularityRankPropertySchema => popularityRankPropertySchema.type === ValueType.Number,
-            `must refer to a "ValueType.Number" property.`,
-          );
-        };
-
         const validateVersionProperty = () => {
           return validateProperty(
             'versionProperty',
@@ -1712,9 +1684,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
         validateUserIdProperty();
         validateGroupIdProperty();
         validateMemberGroupIdProperty();
-        validateAuthorityNormProperty();
-        validatePopularityNormProperty();
-        validatePopularityRankProperty();
         validateVersionProperty();
       })
       .superRefine((data, context) => {
@@ -1744,7 +1713,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
           contextProperties,
           authorityNormProperty,
           popularityNormProperty,
-          popularityRankProperty,
         } = schema.index;
 
         if (authorityNormProperty) {
@@ -1764,16 +1732,6 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
             popularityNormPropertySchema => popularityNormPropertySchema.type === ValueType.Number,
             `must refer to a "ValueType.Number" property.`,
             ['index', 'popularityNormProperty'],
-          );
-        }
-
-        if (popularityRankProperty) {
-          validatePropertyValue(
-            popularityRankProperty,
-            'popularityRankProperty',
-            popularityRankPropertySchema => popularityRankPropertySchema.type === ValueType.Number,
-            `must refer to a "ValueType.Number" property.`,
-            ['index', 'popularityRankProperty'],
           );
         }